MAR-60 fix dynamic route dependend content move dynamic content from layout to dynamic routes
